+++ lib/elastic_apm/instrumenter.rb
@@ -8,10 +8,10 @@ module ElasticAPM
   # rubocop:disable Metrics/ClassLength
   # @api private
   class Instrumenter
-    include Logging
+    TRANSACTION_KEY = :__elastic_instrumenter_transaction_key
+    SPAN_KEY = :__elastic_instrumenter_spans_key
 
-    TRANSACTION_KEY = :__elastic_transaction_key
-    SPAN_KEY = :__elastic_span_key
+    include Logging
 
     # @api private
     class Current
@@ -29,22 +29,24 @@ module ElasticAPM
       end
 
       def spans
-        Thread.current[SPAN_KEY]
+        Thread.current[SPAN_KEY] ||= []
       end
 
       def spans=(spans)
+        Thread.current[SPAN_KEY] ||= []
         Thread.current[SPAN_KEY] = spans
       end
     end
 
-    def initialize(config, &enqueue)
+    def initialize(config, stacktrace_builder:, &enqueue)
       @config = config
+      @stacktrace_builder = stacktrace_builder
       @enqueue = enqueue
 
       @current = Current.new
     end
 
-    attr_reader :config, :enqueue
+    attr_reader :config, :stacktrace_builder, :enqueue
 
     def start
       debug 'Starting instrumenter'
@@ -148,7 +150,8 @@ module ElasticAPM
         transaction_id: transaction.id,
         parent_id: parent.id,
         context: context,
-        trace_context: parent.trace_context
+        trace_context: parent.trace_context,
+        stacktrace_builder: stacktrace_builder
       )
 
       if backtrace && span_frames_min_duration?