+++ spec/elastic_apm/instrumenter_spec.rb
@@ -3,10 +3,17 @@
 module ElasticAPM
   RSpec.describe Instrumenter, :intercept do
     let(:config) { Config.new }
+    let(:stacktrace_builder) { StacktraceBuilder.new(config) }
     let(:callback) { ->(*_) {} }
     before { allow(callback).to receive(:call) }
 
-    subject { Instrumenter.new(config, &callback) }
+    subject do
+      Instrumenter.new(
+        config,
+        stacktrace_builder: stacktrace_builder,
+        &callback
+      )
+    end
 
     context 'life cycle' do
       describe '#stop' do