+++ CHANGELOG.md
@@ -6,9 +6,17 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ## Unreleased
 
+### Added
+
+- Add capture_* options ([#279](https://github.com/elastic/apm-agent-ruby/pull/279))
+
 ### Changed
 
 - Rename `Traceparent` object to `TraceContext` ([#271](https://github.com/elastic/apm-agent-ruby/pull/271))
+
+### Fixed
+
+- An issue where Spans would not get Stacktraces attached ([#282](https://github.com/elastic/apm-agent-ruby/pull/282))
 
 ## 2.1.2 (2018-12-07)